refactor(controller): split tag-library.ts's phases into tag-library/
The tagger CLI is 1274 lines: main()'s orchestration plus four phases
that only main() calls. tag-library.ts keeps main() and the process
lifecycle (664 lines); the phases move out:

  log.ts     the one logging chokepoint (console line + panel event)
  flags.ts   CLI arg parsing and the pre-flight passes
  enrich.ts  phase 0 - Last.fm tags, lyric excerpts, original years
  embed.ts   phase 1 - the text vectors similarity search runs over
  tag.ts     the LLM tagging worker pool (phases 2 and 4)

logEvent earned its own module rather than being passed around: it emits
the terse console line AND the structured sentinel the admin panel reads,
and every phase calls it, so it has to be one shared binding or the two
outputs drift.

Smoke-checked by running the CLI: it parses flags, opens the library DB
and reaches the embedding preflight, which is as far as this box goes
without an Ollama.

No behaviour change. Controller lint clean, all 52 test files pass.
