+++ controller/src/music/tag-library.ts
@@ -35,9 +35,6 @@
 // On boot the library-db auto-migrates any state/moods.json into the SQLite
 // tracks table as legacy v1 entries (see library-db.ts).
 
-import * as subsonic from './subsonic.js';
-import * as lastfm from './lastfm.js';
-import * as musicbrainz from './musicbrainz.js';
 import * as db from './library-db.js';
 import * as settings from '../settings.js';
 import * as embeddings from './embeddings.js';
@@ -45,22 +42,25 @@ import { selectSeeds } from './seed-selector.js';
 import { selectEnrichIds } from './enrich-scope.js';
 import { vote, fuseNeighbours } from './tag-propagator.js';
 import { summariseEval, formatEvalSummary } from './propagation-eval.js';
-import { config } from '../config.js';
-import { loadSecretsIntoEnv } from '../setup/secrets.js';
-import { loadSetupConfig } from '../setup/config.js';
-import { activeModelLabel, primaryLeg, fallbackLeg, probeLegReachable } from '../llm/provider.js';
-import { isUnreachable, isQuotaOrAuthError, errReason } from '../llm/sdk.js';
+import { activeModelLabel } from '../llm/provider.js';
 import { setRawDebugStderrMirror } from '../llm/log.js';
-import { tagBatch, tagOne, taggerBatchSystem, type TagResult } from './tagger-core.js';
+import { taggerBatchSystem } from './tagger-core.js';
 import { runAnalysisPass } from './analyze.js';
-import { reportProgress, formatPhaseBreakdown, sortedPhaseTimings, makeEventLogger } from './tagger-progress.js';
+import { reportProgress, formatPhaseBreakdown, sortedPhaseTimings } from './tagger-progress.js';
 import { planRun } from './rescan-scope.js';
-import { mapPool, memoizeByKey } from '../util/async-pool.js';
 import { acquireStandaloneLock, installPidfileCleanup } from './tagger-lock.js';
+import { phaseEmbed } from './tag-library/embed.js';
+import { logEvent } from './tag-library/log.js';
+import { phaseEnrich } from './tag-library/enrich.js';
+import {
+  applyWizardOverlay,
+  clamp01,
+  parseFlags,
+  reconcileOnly,
+  walkNavidrome,
+} from './tag-library/flags.js';
+import { llmTagInBatches, resolveTagConsumers } from './tag-library/tag.js';
 
-// Emit the terse `[tag] …` console line AND the structured event sentinel the
-// controller relays to the panel — one call site per notable milestone.
-const logEvent = makeEventLogger('tag');
 
 // Close (and TRUNCATE-checkpoint) the library DB on every exit path — this CLI
 // bails via process.exit() from several places, and a bulk run that skips the
@@ -69,173 +69,6 @@ const logEvent = makeEventLogger('tag');
 process.on('exit', () => {
   try { if (db.isOpen()) db.close(); } catch { /* best-effort */ }
 });
-
-// ---------------------------------------------------------------------------
-// CLI arg parsing
-// ---------------------------------------------------------------------------
-
-function parseIntFlag(args: string[], name: string): number | null {
-  const idx = args.indexOf(name);
-  if (idx < 0) return null;
-  const n = parseInt(args[idx + 1], 10);
-  return Number.isFinite(n) ? n : null;
-}
-
-function clamp01(n: number): number {
-  if (!Number.isFinite(n)) return 0;
-  if (n < 0) return 0;
-  if (n > 1) return 1;
-  return n;
-}
-
-interface CliFlags {
-  limit: number;
-  batchSize: number | null;
-  seedCount: number | null;
-  maxRounds: number | null;
-  noPropagate: boolean;
-  reseed: boolean;
-  reEnrich: boolean;
-  skipEnrich: boolean;
-  upgrade: boolean;
-  skipAnalyze: boolean;
-  reAnalyze: boolean;
-  reconcileOnly: boolean;
-  // Skip embed + mood tagging (phases 1-4). Walk, enrich and analyze still run
-  // per their own flags — the admin "Tag moods" step unchecked.
-  skipTag: boolean;
-  // Walk Navidrome but don't prune orphaned rows — the admin "Reconcile with
-  // Navidrome" step unchecked. (A normal run prunes by default.)
-  noPrune: boolean;
-  // Per-run override of the Demucs vocal-activity backfill in Phase 5. --vocal
-  // forces it on, --no-vocal forces it off; neither falls back to the setting
-  // (settings.audio.vocalActivity / ANALYZE_VOCAL_ACTIVITY). The admin Run tab's
-  // "Vocal activity" sub-checkbox drives these so a run can do bpm/key + CLAP
-  // without the slow Demucs pass (or include it) without touching the setting.
-  vocal: boolean;
-  noVocal: boolean;
-  // Re-scan mode (admin Re-scan tab). Fire ONLY the selected re-* passes, each
-  // scoped to already-done tracks; the forward seed→propagate→active-learn
-  // discovery is suppressed so the untagged remainder is never processed. Raw CLI
-  // re-* flags (no --rescan) keep their documented per-flag, full-library meaning.
-  rescan: boolean;
-}
-
-function parseFlags(): CliFlags {
-  const args = process.argv.slice(2);
-  // null = no --batch flag → fall back to settings.embedding.batchSize in main().
-  const rawBatch = parseIntFlag(args, '--batch');
-  return {
-    limit: parseIntFlag(args, '--limit') ?? Infinity,
-    batchSize: rawBatch !== null ? Math.max(1, Math.min(50, rawBatch)) : null,
-    seedCount: parseIntFlag(args, '--seeds'),
-    // null = fall back to settings.embedding.maxActiveLearningRounds
-    maxRounds: parseIntFlag(args, '--max-rounds'),
-    noPropagate: args.includes('--no-propagate'),
-    reseed: args.includes('--reseed'),
-    reEnrich: args.includes('--re-enrich'),
-    skipEnrich: args.includes('--skip-enrich'),
-    // Re-decide moods: re-LLM-tag tagged rows whose prompt/model went stale. Row
-    // selection (db.staleTaggedIds) excludes source='manual' — operator-set tags
-    // are ground truth and never go stale with prompt/model changes.
-    upgrade: args.includes('--upgrade'),
-    skipAnalyze: args.includes('--skip-analyze'),
-    reAnalyze: args.includes('--re-analyze'),
-    // Walk Navidrome and prune library rows for tracks it no longer contains,
-    // then exit — no embeddings, no LLM. The admin "Reconcile with Navidrome"
-    // button drives this so orphaned entries can be cleared without paying for
-    // a full tag/analyze pass (works even at 100% coverage).
-    reconcileOnly: args.includes('--reconcile-only'),
-    skipTag: args.includes('--skip-tag'),
-    noPrune: args.includes('--no-prune'),
-    vocal: args.includes('--vocal'),
-    noVocal: args.includes('--no-vocal'),
-    rescan: args.includes('--rescan'),
-  };
-}
-
-// Walk the entire Navidrome catalogue, upserting each song's metadata into the
-// tracks table and collecting the live id set. Shared by the full tagger run
-// (Phase A) and the standalone --reconcile-only path. Cheap: metadata only, no
-// embeddings or LLM calls.
-async function walkNavidrome(): Promise<{ walked: number; liveIds: Set<string> }> {
-  reportProgress({ phase: 'walk', label: 'Scanning Navidrome library', done: 0 });
-  let walked = 0;
-  const liveIds = new Set<string>();
-  for await (const song of subsonic.iterateAllSongs()) {
-    db.upsertTrackMeta(song.id, {
-      title: song.title,
-      artist: song.artist,
-      album: song.album,
-      year: song.year,
-      // Album-level era signals (issue #842). The album's originalReleaseDate
-      // is each track's original year ONLY on a non-compilation album — a
-      // compilation's original date is the compilation's own, so its tracks
-      // stay unresolved here and phase-0 asks MusicBrainz per track instead.
-      originalYear: song.albumIsCompilation ? null : song.albumOriginalYear ?? null,
-      isCompilation: song.albumIsCompilation ?? null,
-      genres: subsonic.songGenres(song),
-      duration: song.duration,
-    });
-    liveIds.add(song.id);
-    walked += 1;
-    if (walked % 500 === 0) {
-      console.log(`[tag] walked ${walked} tracks`);
-      reportProgress({ phase: 'walk', label: 'Scanning Navidrome library', done: walked });
-    }
-  }
-  logEvent('info', `Scanned ${walked.toLocaleString('en-GB')} tracks`);
-  return { walked, liveIds };
-}
-
-// Standalone reconcile: diff library-db against the live Navidrome catalogue and
-// drop rows (and their vectors) for tracks that are gone. No embedding preflight
-// and no LLM — opens the existing DB at its stored dim so vectors are untouched.
-async function reconcileOnly() {
-  await db.open({ embeddingDim: embeddings.resolveEmbeddingDim(), adoptStoredDim: true });
-  console.log('[tag] reconcile-only: walking Navidrome to prune orphaned rows');
-  const { walked, liveIds } = await walkNavidrome();
-  let pruned = 0;
-  if (walked > 0) {
-    pruned = db.pruneMissingTracks(liveIds);
-    console.log(`[tag] reconcile pruned ${pruned} orphaned tracks no longer in Navidrome`);
-  } else {
-    // A transient empty Navidrome response must never wipe the DB.
-    console.warn('[tag] reconcile: Navidrome returned 0 tracks — skipping prune');
-  }
-  reportProgress({
-    phase: 'done',
-    label: pruned > 0
-      ? `Removed ${pruned} track${pruned === 1 ? '' : 's'} no longer in Navidrome`
-      : 'Library is in sync with Navidrome',
-    done: pruned,
-  });
-  console.log(`[tag] reconcile complete (walked ${walked}, pruned ${pruned})`);
-  process.exit(0);
-}
-
-// Mirrors server.ts boot: cloud API keys from secrets.env, Navidrome creds
-// from setup-config.json. Standalone CLIs skip server.ts, so without this
-// they fall back to the hardcoded `http://navidrome:4533` and ENOTFOUND on
-// any install with a custom Navidrome host.
-async function applyWizardOverlay() {
-  try {
-    await loadSecretsIntoEnv();
-  } catch (err: any) {
-    console.error('[secrets] load failed:', err.message);
-  }
-  try {
-    const sc = await loadSetupConfig();
-    if (sc.navidrome) {
-      if (!process.env.NAVIDROME_URL && sc.navidrome.url) config.navidrome.url = sc.navidrome.url;
-      if (!process.env.NAVIDROME_USER && sc.navidrome.user) config.navidrome.user = sc.navidrome.user;
-      if (!process.env.NAVIDROME_PASS && sc.navidrome.pass)
-        config.navidrome.password = sc.navidrome.pass;
-    }
-  } catch (err: any) {
-    console.error('[setup-config] load failed:', err.message);
-  }
-}
 
 // ---------------------------------------------------------------------------
 // Main orchestration
@@ -823,448 +656,6 @@ function finish(
     `Library now: ${(s.total ?? 0).toLocaleString('en-GB')} tagged · ${moods} moods · ${genres} genres · ` +
       `${(s.withEmbedding ?? 0).toLocaleString('en-GB')} embedded${src ? ` · ${src}` : ''}`,
   );
-}
-
-// ---------------------------------------------------------------------------
-// Phase 0 — Enrichment (Last.fm tags + lyric excerpts)
-// ---------------------------------------------------------------------------
-
-async function phaseEnrich(ids: string[], reEnrich: boolean): Promise<void> {
-  const enrichCfg = (settings.get() as any).embedding?.enrichment ?? {};
-  // A configured Last.fm api_key (LASTFM_API_KEY / scrobble.lastfm.apiKey) lets
-  // us hit the Last.fm API directly (music/lastfm.ts), which actually returns
-  // tag[]. The tri-state gate (shared with the retag route via
-  // lastfm.lastfmEnrichEnabled) keeps keyless vanilla-Navidrome installs from
-  // wasting a round trip per artist on the tag-less getArtistInfo2 path.
-  const hasKey = lastfm.hasLastfmKey();
-  const lastfmEnabled = lastfm.lastfmEnrichEnabled(enrichCfg.lastfmTags, hasKey);
-  const lyricsEnabled = enrichCfg.lyrics !== false;
-  // Original-year resolution for compilation-album tracks (issue #842) —
-  // keyless MusicBrainz, so default-on; disable via
-  // settings.embedding.enrichment.originalYear = false.
-  const originalYearEnabled = enrichCfg.originalYear !== false;
-  // The original-year backfill scopes itself in SQL (compilation tracks with
-  // no resolved year), NOT by the tagger's untagged/enriched id set — the
-  // column landed after most libraries were tagged, and gating on the enrich
-  // scope (or enrichedAt) would never backfill an already-tagged library.
-  const yearIds = originalYearEnabled ? db.idsNeedingOriginalYear(reEnrich) : [];
-  if (ids.length === 0 && yearIds.length === 0) return;
-  if (!lastfmEnabled && !lyricsEnabled && yearIds.length === 0) {
-    console.log('[tag] phase-0 skipped: lastfmTags and lyrics disabled, no original-year lookups pending');
-    return;
-  }
-  if (lastfmEnabled) {
-    console.log(
-      `[tag] phase-0 Last.fm tags via ${hasKey ? 'direct API (artist.getTopTags)' : 'Navidrome getArtistInfo2 (no api_key — likely empty)'}`,
-    );
-  }
-  reportProgress({ phase: 'enrich', label: 'Enriching metadata', done: 0, total: ids.length });
-  // Per-artist Last.fm dedup: memoize on the in-flight PROMISE (not the resolved
-  // value) so the concurrent pool below shares one API call across every track by
-  // the same artist. Direct Last.fm API when a key is present (returns crowd tags
-  // on vanilla Navidrome), else Navidrome's getArtistInfo2 — shared with the
-  // single-track retag route so the two can't drift (see lastfm.ts).
-  const artistTags = memoizeByKey<string[]>(artist =>
-    lastfm.getArtistTags(artist, { count: 10 }).then(t => t ?? []).catch(() => []),
-  );
-
-  let enrichedTracks = 0;
-  let enrichedLyrics = 0;
-  let enrichedTags = 0;
-  let enrichedYears = 0;
-
-  // Enrichment is I/O-bound (Last.fm + Navidrome lyrics fetches), so a serial
-  // await-per-track loop spends almost all its time blocked on the network — on a
-  // large library that's the difference between minutes and an hour. Drain the id
-  // list with a bounded pool instead. DB writes go through better-sqlite3
-  // (synchronous), so they're naturally serialised on the single-threaded event
-  // loop between awaits — no locking needed. Pool size is gentle by default (6)
-  // and tunable via TAG_ENRICH_CONCURRENCY so a small Navidrome / Last.fm budget
-  // can dial it down.
-  const concurrency = Math.max(
-    1,
-    Math.min(32, parseInt(process.env.TAG_ENRICH_CONCURRENCY || '', 10) || 6),
-  );
-
-  // Skip the per-track Last.fm/lyrics loop entirely when both are disabled —
-  // running it anyway would stamp enrichedAt with empty rows and mask a later
-  // re-enable. The original-year loop below has its own scope + skip stamp.
-  const metaIds = lastfmEnabled || lyricsEnabled ? ids : [];
-  await mapPool(metaIds, concurrency, async (id) => {
-    const t = db.getTrack(id);
-    if (!t) return;
-    if (!reEnrich && t.enrichedAt) return;
-
-    let lastfmTags: string[] | null = null;
-    if (lastfmEnabled && t.artist) {
-      const tags = await artistTags(t.artist);
-      lastfmTags = tags.length ? tags : null;
-    }
-
-    let lyricExcerpt: string | null = null;
-    if (lyricsEnabled) {
-      try {
-        const raw = await subsonic.getLyrics(id);
-        if (typeof raw === 'string' && raw.trim()) {
-          lyricExcerpt = raw.trim();
-        }
-      } catch { /* ignore */ }
-    }
-
-    db.upsertTrackEnrichment(id, { lastfmTags, lyricExcerpt });
-    enrichedTracks += 1;
-    if (lastfmTags && lastfmTags.length) enrichedTags += 1;
-    if (lyricExcerpt) enrichedLyrics += 1;
-    if (enrichedTracks % 100 === 0) {
-      console.log(
-        `[tag] enriched ${enrichedTracks}/${metaIds.length} (lastfm: ${enrichedTags}, lyrics: ${enrichedLyrics})`,
-      );
-      reportProgress({ phase: 'enrich', label: 'Enriching metadata', done: enrichedTracks, total: metaIds.length });
-    }
-  });
-
-  // Original-year backfill (issue #842): compilation-album tracks whose plain
-  // `year` is the compilation's own release date. Resolved via MusicBrainz
-  // (music/musicbrainz.ts — earliest first-release-date across matching
-  // recordings, exact-MBID first). Effectively serial regardless of pool width
-  // (the client's 1 req/s throttle), so a big compilation-heavy library takes
-  // a while on the first pass — hence the per-track checked_at stamp (hit OR
-  // miss) that makes every later pass skip straight past it. The per-song
-  // Navidrome fetch just recovers the recording MBID (not stored in
-  // library.db); it's dwarfed by the MB throttle.
-  if (yearIds.length) {
-    console.log(`[tag] phase-0 original years: ${yearIds.length} compilation tracks to resolve via MusicBrainz (~1/s)`);
-    reportProgress({ phase: 'enrich', label: 'Resolving original years', done: 0, total: yearIds.length });
-    let checkedYears = 0;
-    await mapPool(yearIds, Math.min(concurrency, 4), async (id) => {
-      const t = db.getTrack(id);
-      if (!t || !musicbrainz.needsOriginalYearLookup(t, reEnrich)) return;
-      let mbid: string | null = null;
-      try {
-        mbid = (await subsonic.getSong(id))?.musicBrainzId || null;
-      } catch { /* MBID is optional — the search path covers it */ }
-      const year = await musicbrainz.lookupOriginalYear({ title: t.title, artist: t.artist, mbid, year: t.year });
-      db.setOriginalYear(id, year);
-      if (year != null) enrichedYears += 1;
-      checkedYears += 1;
-      if (checkedYears % 25 === 0) {
-        console.log(`[tag] original years ${checkedYears}/${yearIds.length} (${enrichedYears} resolved)`);
-        reportProgress({ phase: 'enrich', label: 'Resolving original years', done: checkedYears, total: yearIds.length });
-      }
-    });
-  }
-
-  logEvent(
-    'info',
-    `Metadata fetched for ${enrichedTracks.toLocaleString('en-GB')} tracks ` +
-      `(${enrichedTags} Last.fm, ${enrichedLyrics} lyrics, ${enrichedYears} original years)`,
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Phase 1 — Embed
-// ---------------------------------------------------------------------------
-
-async function phaseEmbed(
-  targetIds: string[],
-  batchSize: number,
-  // The index's task-prefix mode (resolved once in run()) — every document
-  // this phase writes must match the vectors already in the index.
-  textMode: embeddings.IndexTextMode,
-): Promise<void> {
-  // Embed any track in scope that doesn't already have a vector. Includes
-  // already-tagged tracks (legacy v1) so they can serve as KNN neighbours.
-  const needsEmbed: string[] = [];
-  for (const id of targetIds) {
-    if (!db.hasVector(id)) needsEmbed.push(id);
-  }
-  // Also embed all already-tagged tracks that don't have vectors yet (legacy
-  // v1 imports). Without this they can't anchor the KNN graph.
-  for (const id of db.allTaggedIds()) {
-    if (!db.hasVector(id)) needsEmbed.push(id);
-  }
-  // Dedup
-  const unique = [...new Set(needsEmbed)];
-  if (unique.length === 0) {
-    console.log('[tag] phase-1 nothing to embed');
-    return;
-  }
-  logEvent('info', `Building similarity vectors for ${unique.length.toLocaleString('en-GB')} tracks…`);
-  reportProgress({ phase: 'embed', label: 'Embedding tracks', done: 0, total: unique.length });
-
-  const embedBatchSize = Math.max(8, Math.min(64, batchSize * 2));
-  for (let i = 0; i < unique.length; i += embedBatchSize) {
-    const batch = unique.slice(i, i + embedBatchSize);
-    const songs = batch.map(id => db.getTrack(id)).filter((t): t is db.TrackRecord => !!t);
-    const texts = songs.map(t =>
-      embeddings.formatTrackText(
-        { title: t.title, artist: t.artist, album: t.album, year: t.year, genres: t.genres },
-        { lastfmTags: t.lastfmTags, lyricExcerpt: t.lyricExcerpt },
-      ),
-    );
-    let vecs: number[][];
-    try {
-      vecs = await embeddings.embedDocTexts(texts, textMode);
-    } catch (err: any) {
-      console.error(`[tag] embedding batch failed at offset ${i}: ${err.message}`);
-      throw err;
-    }
-    for (let j = 0; j < songs.length; j++) {
-      db.upsertTrackVector(songs[j].id, vecs[j]);
-    }
-    if ((i + batch.length) % 500 === 0 || i + batch.length === unique.length) {
-      console.log(`[tag] embedded ${i + batch.length}/${unique.length}`);
-      reportProgress({ phase: 'embed', label: 'Embedding tracks', done: i + batch.length, total: unique.length });
-    }
-  }
-}
-
-// ---------------------------------------------------------------------------
-// LLM tagging helper (reused by phase 2 + phase 4)
-// ---------------------------------------------------------------------------
-
-// A single LLM worker the batch loop pulls through. `pin` selects which leg
-// each call targets (undefined → normal primary→fallback failover, used in
-// single-LLM mode); `label` is stamped on every track this consumer tags so the
-// per-track provenance is honest across two different models (discussion #320).
-interface TagConsumer {
-  pin?: 'primary' | 'fallback';
-  label: string;
-}
-
-interface TagState {
-  tagged: number;
-  callCount: number;
-  processed: number;
-  // Tracks that came back null from the LLM (batch entry dropped / per-track
-  // salvage failed) — surfaced in the progress channel.
-  errors: number;
-  byLeg: Record<string, number>;
-}
-
-// Which pipeline phase a runConsumer() call is tagging for — only used to
-// stamp the progress channel ('seed' = phase 2, 'learn' = phase 4 rounds).
-interface TagPhaseInfo {
-  phase: 'seed' | 'learn';
-  round?: number;
-}
-
-// Tag one batch with one consumer's leg. Returns the count actually tagged.
-// Throws ONLY when a pinned leg's host is unreachable — the caller requeues the
-// whole batch and drops the consumer. Upserts happen only after the batch fully
-// resolves, so a rethrow mid-batch persists nothing: the requeue is lossless.
-// Non-unreachable failures (small models dropping list entries) salvage per
-// track exactly as before, so one bad line never sinks 25 tracks.
-async function processBatch(
-  batch: string[],
-  consumer: TagConsumer,
-  promptHash: string,
-  source: db.TagSource,
-  state: TagState,
-): Promise<number> {
-  const songs = batch.map(id => db.getTrack(id)).filter((t): t is db.TrackRecord => !!t);
-  if (songs.length === 0) return 0;
-  const input = songs.map(t => ({
-    title: t.title ?? undefined,
-    artist: t.artist ?? undefined,
-    album: t.album ?? undefined,
-    year: t.year ?? undefined,
-    genres: t.genres,
-  }));
-  const opts = consumer.pin ? { leg: consumer.pin } : {};
-
-  let results: Array<TagResult | null>;
-  try {
-    results = await tagBatch(input, opts);
-    state.callCount += 1;
-  } catch (err: any) {
-    // A pinned leg that can't recover this run — host down, OR a
-    // quota/usage-limit/auth rejection (#438): rethrow BEFORE the per-track
-    // salvage, otherwise we'd grind 25 serial connect-timeouts (or 25 identical
-    // 429s) against a leg that won't answer. The surviving consumer redoes the
-    // requeued batch.
-    if (consumer.pin && (isUnreachable(err) || isQuotaOrAuthError(err))) throw err;
-    // A "batch length mismatch" is NOT a failure. Some models (e.g. Mercury, and
-    // small local models) don't return one structured-output entry per input
-    // track, so we tag each track individually this batch — same seed set, same
-    // cost envelope (only the seeds ever hit the LLM), just slower. Log it as an
-    // expected degrade, not an error, so it doesn't read as something broken.
-    // Genuine batch errors keep the error-level line with their message.
-    const perTrackDegrade = /batch length mismatch/i.test(err.message || '');
-    if (perTrackDegrade) {
-      logEvent(
-        'warning',
-        `${consumer.label} didn't return one entry per track — tagging ` +
-          `${songs.length} tracks individually this batch (expected for some models; just slower)`,
-      );
-    } else {
-      logEvent(
-        'error',
-        `LLM batch failed (${songs.length} tracks) on ${consumer.label}: ${err.message} — falling back to per-track`,
-      );
-    }
-    results = [];
-    for (const song of input) {
-      try {
-        results.push(await tagOne(song, opts));
-        state.callCount += 1;
-      } catch (oneErr: any) {
-        // Leg unusable mid-salvage (host died, or quota/auth) — bail the whole
-        // batch (nothing upserted yet).
-        if (consumer.pin && (isUnreachable(oneErr) || isQuotaOrAuthError(oneErr))) throw oneErr;
-        console.error(`[tag] per-track tag failed on ${consumer.label}: ${oneErr.message}`);
-        results.push(null);
-      }
-    }
-  }
-
-  let tagged = 0;
-  for (let j = 0; j < songs.length; j++) {
-    const result = results[j];
-    if (!result) {
-      state.errors += 1;
-      continue;
-    }
-    const { moods, energy } = result;
-    db.upsertTrackTags(songs[j].id, {
-      moods,
-      energy,
-      source,
-      confidence: null,
-      promptHash,
-      model: consumer.label,
-    });
-    tagged += 1;
-  }
-  return tagged;
-}
-
-// Drain the shared `batches` queue with one consumer. `shift()` between awaits
-// is atomic (single-threaded event loop), so two consumers never pull the same
-// batch. In dual mode a pinned consumer whose host dies requeues its batch and
-// returns; `onDrop` reports how many legs remain.
-async function runConsumer(
-  batches: string[][],
-  consumer: TagConsumer,
-  promptHash: string,
-  source: db.TagSource,
-  total: number,
-  state: TagState,
-  phaseInfo: TagPhaseInfo,
-  onDrop: ((err: any) => number) | null,
-): Promise<void> {
-  for (;;) {
-    const batch = batches.shift();
-    if (!batch) return;
-    try {
-      const n = await processBatch(batch, consumer, promptHash, source, state);
-      state.tagged += n;
-      state.byLeg[consumer.label] = (state.byLeg[consumer.label] || 0) + n;
-    } catch (err: any) {
-      // processBatch rethrows only when a pinned leg can't recover this run:
-      // the host is down (isUnreachable) OR the provider refused the leg with a
-      // quota / credit / usage-limit / auth error (isQuotaOrAuthError, #438).
-      // Name the real reason — logging every drop as "unreachable" misdirected an
-      // operator whose OpenRouter credits had simply run out (Discord).
-      batches.unshift(batch);
-      const remaining = onDrop ? onDrop(err) : 0;
-      const reason = isQuotaOrAuthError(err) ? 'quota/credit/auth rejected' : 'host unreachable';
-      logEvent(
-        'error',
-        `LLM leg ${consumer.label} dropped — ${reason}: ${errReason(err)} (${remaining} leg(s) left)`,
-      );
-      return;
-    }
-    state.processed += 1;
-    if (state.processed % 4 === 0) {
-      console.log(`[tag] LLM-tagged ${state.tagged}/${total}`);
-    }
-    reportProgress({
-      phase: phaseInfo.phase,
-      label: 'Tagging with LLM',
-      done: state.tagged,
-      total,
-      round: phaseInfo.round,
-      errors: state.errors || undefined,
-      llm: { legs: state.byLeg },
-    });
-  }
-}
-
-// Phase 2 + phase 4 LLM tagging. One consumer (single-LLM mode, failover-capable
-// calls) or two (dual-LLM mode, one pinned per leg) drain a shared batch queue.
-async function llmTagInBatches(
-  ids: string[],
-  batchSize: number,
-  promptHash: string,
-  source: db.TagSource,
-  consumers: TagConsumer[],
-  phaseInfo: TagPhaseInfo,
-): Promise<{ tagged: number; callCount: number; byLeg: Record<string, number> }> {
-  const batches: string[][] = [];
-  for (let i = 0; i < ids.length; i += batchSize) batches.push(ids.slice(i, i + batchSize));
-
-  const state: TagState = { tagged: 0, callCount: 0, processed: 0, errors: 0, byLeg: {} };
-  reportProgress({
-    phase: phaseInfo.phase,
-    label: 'Tagging with LLM',
-    done: 0,
-    total: ids.length,
-    round: phaseInfo.round,
-  });
-
-  if (consumers.length <= 1) {
-    // Single consumer — no requeue/drop; the unpinned call already fails over
-    // internally, so an error means the batch is genuinely unworkable this run.
-    await runConsumer(batches, consumers[0], promptHash, source, ids.length, state, phaseInfo, null);
-  } else {
-    let alive = consumers.length;
-    let quotaOrAuthDrop = false;
-    await Promise.all(
-      consumers.map(c =>
-        runConsumer(batches, c, promptHash, source, ids.length, state, phaseInfo, (err: any) => {
-          if (isQuotaOrAuthError(err)) quotaOrAuthDrop = true;
-          return --alive;
-        })),
-    );
-    if (batches.length > 0) {
-      const abandoned = batches.reduce((n, b) => n + b.length, 0);
-      const hint = quotaOrAuthDrop
-        ? ' — a leg was refused for quota/credit/auth; check the provider credit balance, spend cap, or API key'
-        : '';
-      logEvent('warning', `All LLM legs dropped — ${abandoned} tracks left for next run${hint}`);
-    }
-  }
-  return { tagged: state.tagged, callCount: state.callCount, byLeg: state.byLeg };
-}
-
-// Decide the LLM consumers for this run. Dual-LLM mode activates automatically
-// when a fallback is configured, distinct from the primary, and its host answers
-// a cheap probe — then both boxes tag in parallel off a shared queue. Otherwise a
-// single failover-capable consumer (discussion #320).
-async function resolveTagConsumers(): Promise<TagConsumer[]> {
-  const primary = primaryLeg();
-  const fb = fallbackLeg();
-  if (!fb) return [{ label: primary.label }];
-
-  const sameHost =
-    (primary.cfg.ollamaUrl || '') === (fb.cfg.ollamaUrl || '') &&
-    (primary.cfg.baseUrl || '') === (fb.cfg.baseUrl || '');
-  if (fb.label === primary.label && sameHost) {
-    logEvent('info', 'Fallback LLM identical to primary — single-LLM mode');
-    return [{ label: primary.label }];
-  }
-
-  if (!(await probeLegReachable(fb))) {
-    logEvent('info', `Fallback LLM (${fb.label}) unreachable — single-LLM mode`);
-    return [{ label: primary.label }];
-  }
-
-  logEvent('info', `Dual-LLM mode active: primary=${primary.label} + fallback=${fb.label}`);
-  return [
-    { pin: 'primary', label: primary.label },
-    { pin: 'fallback', label: fb.label },
-  ];
 }
 
 // Explicitly exit on success. The local analyze backend (analyzer.ts) is a
